requirements.txt
alembic==1.7.3
blueprint==3.4.2
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
environs==9.3.4
Flask==2.0.2
Flask-HTTPAuth==4.4.0
Flask-JWT-Extended==4.3.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==2.0.1
